+++ app/playlists_ui/playlistsmodel.h
@@ -19,6 +19,12 @@ namespace PlaylistsUi {
     int rowCount(const QModelIndex &parent = QModelIndex()) const override;
     QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
 
+    Qt::ItemFlags flags(const QModelIndex &index) const override;
+    Qt::DropActions supportedDropActions() const override;
+    QStringList mimeTypes() const override;
+    QMimeData *mimeData(const QModelIndexList &indexes) const override;
+    bool dropMimeData(const QMimeData *data, Qt::DropAction action, int row, int column, const QModelIndex &parent) override;
+
     QModelIndex buildIndex(int row) const;
     QModelIndex append(std::shared_ptr<Playlist::Playlist> item);
     virtual void remove(const QModelIndex &index);